api: normalize SDK identity parameters to strings at request intake
app_key, device_id and old_device_id are always strings by SDK/API
contract. Previously each consumer cast them ad hoc (some lookups cast,
some did not), which is easy to get wrong and led to inconsistent
handling between the write and fetch paths.

Normalize these parameters to strings once, centrally, in processRequest
right after the query string and body are merged into params.qstring,
and drop the scattered per-handler/per-query casts that this makes
redundant. Bulk sub-requests carry their own per-item app_key/device_id
(not the top-level qstring) and keep their existing normalization in
processBulkRequest.

Adds a regression test asserting the SDK fetch path treats app_key
strictly as a string and does not match an app by prefix.
