+++ api/utils/requestProcessor.js
@@ -134,6 +134,21 @@ const processRequest = (params) => {
         }
     }
 
+    //Normalize known scalar identity parameters to strings before they reach
+    //any database lookup. These are always strings by SDK/API contract, so a
+    //non-string value (for example an object) would only ever be an attempt to
+    //pass a query operator into a lookup. This is the single authoritative
+    //place that normalizes them, so individual handlers and queries must not
+    //re-cast. Bulk sub-requests carry their own per-item app_key/device_id
+    //(not the top-level qstring) and are normalized separately in
+    //processBulkRequest.
+    var stringParams = ["app_key", "device_id", "old_device_id"];
+    for (let s = 0; s < stringParams.length; s++) {
+        if (typeof params.qstring[stringParams[s]] !== "undefined" && params.qstring[stringParams[s]] !== null) {
+            params.qstring[stringParams[s]] = params.qstring[stringParams[s]] + "";
+        }
+    }
+
     if (params.qstring.app_id && params.qstring.app_id.length !== 24) {
         common.returnMessage(params, 400, 'Invalid parameter "app_id"');
         return false;
@@ -1543,10 +1558,8 @@ const processRequest = (params) => {
                     return false;
                 }
                 else {
-                    //make sure device_id is string
-                    params.qstring.device_id += "";
-                    params.qstring.app_key += "";
                     // Set app_user_id that is unique for each user of an application.
+                    // app_key/device_id are already normalized to strings in processRequest.
                     params.app_user_id = common.crypto.createHash('sha1')
                         .update(params.qstring.app_key + params.qstring.device_id + "")
                         .digest('hex');
@@ -2976,7 +2989,7 @@ const processRequest = (params) => {
                     return false;
                 }
                 else {
-                    params.qstring.device_id += "";
+                    // app_key/device_id are already normalized to strings in processRequest.
                     params.app_user_id = common.crypto.createHash('sha1')
                         .update(params.qstring.app_key + params.qstring.device_id + "")
                         .digest('hex');
@@ -2999,7 +3012,7 @@ const processRequest = (params) => {
                     return false;
                 }
                 else {
-                    params.qstring.device_id += "";
+                    // app_key/device_id are already normalized to strings in processRequest.
                     params.app_user_id = common.crypto.createHash('sha1')
                         .update(params.qstring.app_key + params.qstring.device_id + "")
                         .digest('hex');
@@ -3482,7 +3495,7 @@ const validateAppForWriteAPI = (params, done, try_times) => {
         return done ? done() : false;
     }
 
-    common.readBatcher.getOne("apps", {'key': params.qstring.app_key + ""}, (err, app) => {
+    common.readBatcher.getOne("apps", {'key': params.qstring.app_key}, (err, app) => {
         if (!app) {
             common.returnMessage(params, 400, 'App does not exist');
             params.cancelRequest = "App not found or no Database connection";
@@ -3627,7 +3640,7 @@ const validateAppForFetchAPI = (params, done, try_times) => {
     if (ignorePossibleDevices(params)) {
         return done ? done() : false;
     }
-    common.readBatcher.getOne("apps", {'key': params.qstring.app_key + ""}, (err, app) => {
+    common.readBatcher.getOne("apps", {'key': params.qstring.app_key}, (err, app) => {
         if (!app) {
             common.returnMessage(params, 400, 'App does not exist');
             params.cancelRequest = "App not found or no Database connection";
